fix(dashboard): improve data accuracy with date validation, gap filling, and Numeric precision
- Add date swap when d_from > d_to in dashboard and category API endpoints
- Fill missing days in daily trend chart so x-axis has no gaps
- Link monthly chart to filter date range instead of fixed last 3 months
- Change Float columns to Numeric(12,2) / Numeric(15,6) for financial precision
- Remove hardcoded "(Last 4 Months)" labels from dashboard charts
- Add migration for Float to Numeric column type changes
- Add tests for date swap, daily gap filling, monthly filter, and API date swap

+++ models/cashflow.py
@@ -16,7 +16,7 @@ class CashflowTransaction(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     date = db.Column(db.Date, nullable=False)
     type = db.Column(db.String(10), nullable=False)  # 'income' veya 'expense'
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(12, 2), nullable=False)
     description = db.Column(db.Text)
     category_id = db.Column(db.Integer, db.ForeignKey('category.id'), nullable=False)
     external_transaction_id = db.Column(db.String(255), nullable=True, index=True)
